Gating tests
This commit gets rid of the existing, unused, nonworking
and nonmaintained tests. The person who created them
is (AFAICT) no longer with Red Hat; and the tests weren't
being run in gating; and they aren't even in working
condition.

Replace with our standard convention: install the -tests
package, and run it with BATS. This setup is more complex
than podman/skopeo because we need to fire up a local
registry before the tests begin, then tear it down at
the end.

I am expecting these tests to fail in actual gating due
to what I believe are actual buildah problems.

Signed-off-by: Ed Santiago <santiago@redhat.com>